Update GitHub Actions deployment workflow
The workflow builds and deploys the blog to production.
However, I have noticed that the Jekyll version used locally
and during the build differs.

In the job logs we can find:

> Configuration file: /github/workspace/./_config.yml
>       GitHub Pages: github-pages v228
>       GitHub Pages: jekyll v3.9.3

This is because GitHub still uses the old Jekyll.
This is a feature of the actions/jekyll-build-pages [1] action.

> A simple GitHub Action for producing Jekyll build artifacts
> compatible with GitHub Pages.

Since we want to use a newer version of Jekyll,
we need to install it from Gemfile.

[1]: https://github.com/actions/jekyll-build-pages