Add missing closing tag for span
Modern browsers have an error correction system.
Therefore, when the HTML code is not completely correct,
they deal with the correct interpretation.
However, the Markdown engine does not have such a mechanism.
If the HTML is written incorrectly,
Markdown will try to process it in the own way.

In the previous commit I changed the HTML file to Markdown.
But after starting the blog locally,
I noticed that the tags page looked different.
Incorrect tags have been escaped.

It turned out that the closing tag was missing.
Without it, Markdown treated other closing tags as quotes
and escaped them.

Here is an example:

```
</time>&lt;/a&gt;&lt;/li&gt;
```

After making the correction, everything works as it should.