Style the pre and code tag for poetry
This is a lifestyle blog.
So there is no code in the articles.
That's why I decided to use tags intended for code for poetry.
By poetry I mean poems, but also lyrics of songs.
Poetry is characterized by the fact
that the division into lines is important.
Therefore, enters should be interpreted literally.
This is somewhat similar to code.