Change a blockquote from HTML to Markdown
Markdown linter reported  that the following rules were broken:

> MD033/no-inline-html Inline HTML [Element: blockquote]
> MD033/no-inline-html Inline HTML [Element: p]
> MD033/no-inline-html Inline HTML [Element: em]
> MD033/no-inline-html Inline HTML [Element: footer]

This is because the quote was inserted into the article using HTML.
Since we have recently introduced changes to CSS,
we can replace this HTML code with regular Markdown.
Thanks to this, the visual effect will be the same.
And at the same time,
we maintain the consistency of using Markdown syntax.